Bug excepciones con MongoDBLegacy y MongoDB

+++ app/model/ValueFormater.php
@@ -1,10 +1,7 @@
 <?php
-
 namespace Acd\Model;
 
-class ValueFormaterInvalidFormatException extends \exception
-{
-}
+class ValueFormaterInvalidFormatException extends \exception {}
 class ValueFormater
 {
 	const TYPE_TEXT_SIMPLE = 'text_simple';
@@ -31,8 +28,7 @@ class ValueFormater
 	const PERIOD_OF_VALIDITY_START = 'start';
 	const PERIOD_OF_VALIDITY_END = 'end';
 
-	public static function decode($value, $type, $format)
-	{
+	public static function decode($value, $type, $format) {
 		//throw new StorageKeyInvalidException("Invalid format type $format.");
 
 		// Array of procesing functions
@@ -42,7 +38,8 @@ class ValueFormater
 				$valueDecode = \DateTime::createFromFormat('Y-m-d', $value);
 				$valueDecode->setTime(0, 0, 0);
 				return $valueDecode->getTimeStamp();
-			} else {
+			}
+			else {
 				return '';
 			}
 		};
@@ -52,9 +49,9 @@ class ValueFormater
 				ValueFormater::PERIOD_OF_VALIDITY_START => -INF,
 				ValueFormater::PERIOD_OF_VALIDITY_END => INF,
 			];
-			if (is_array($aValue)) {
+			if (is_array($aValue)){
 				foreach ($aValue as $attributeName => $value) {
-					if ($value) {
+					if($value) {
 						$valueDecode = \DateTime::createFromFormat('Y-m-d', $value);
 						$valueDecode->setTime(0, 0, 0);
 						$result[$attributeName] = $valueDecode->getTimeStamp();
@@ -68,7 +65,8 @@ class ValueFormater
 			if ($value) {
 				$valueDecode = \DateTime::createFromFormat('Y-m-d*H:i:s*', $value);
 				return $valueDecode->getTimeStamp();
-			} else {
+			}
+			else {
 				return '';
 			}
 		};
@@ -78,9 +76,9 @@ class ValueFormater
 				ValueFormater::PERIOD_OF_VALIDITY_END => INF,
 			];
 
-			if (is_array($aValue)) {
+			if (is_array($aValue)){
 				foreach ($aValue as $attributeName => $value) {
-					if ($value) {
+					if($value) {
 						$valueDecode = \DateTime::createFromFormat('Y-m-d*H:i:s*', $value);
 						$result[$attributeName] = $valueDecode->getTimeStamp();
 					}
@@ -90,11 +88,12 @@ class ValueFormater
 			return $result;
 		};
 		$formater[self::TYPE_TAGS][self::FORMAT_EDITOR] = function ($value) {
-			if ($value) {
+			if($value) {
 				$value = trim($value);
 				$value = preg_replace('/\s+/', ' ', $value);
 				return explode(',', $value);
-			} else {
+			}
+			else {
 				return array();
 			}
 		};
@@ -102,9 +101,10 @@ class ValueFormater
 			return $value == 1;
 		};
 		$formater[self::TYPE_LIST_MULTIPLE][self::FORMAT_EDITOR] = function ($value) {
-			if ($value) {
+			if($value) {
 				return $value;
-			} else {
+			}
+			else {
 				return array();
 			}
 		};
@@ -112,7 +112,8 @@ class ValueFormater
 			// cleanHtmlFragment
 			if (extension_loaded('tidy') === true) {
 				$encoding = 'utf8';
-				$tidy_config = array(
+				$tidy_config = array
+				(
 					//'anchor-as-name' => false, //error en producción
 					'break-before-br' => true,
 					'char-encoding' => $encoding,
@@ -160,7 +161,8 @@ class ValueFormater
 				$tidy->cleanRepair();
 				$result = \tidy_get_output($tidy);
 				return $result;
-			} else {
+			}
+			else {
 				throw new ValueFormaterInvalidFormatException("Text not validated. Tidy extension not instaled", 1);
 				return $value;
 			}
@@ -179,7 +181,8 @@ class ValueFormater
 			// val[rgb] hex format, ej. #46e7da
 			if (isset($value['empty']) && $value['empty']) {
 				$result = null;
-			} else {
+			}
+			else {
 				$result = $value['rgb'];
 			}
 			return $result;
@@ -188,7 +191,8 @@ class ValueFormater
 			// val[rgb] + val[alfa] y hex format, ej. #46e7da01
 			if (isset($value['empty']) && $value['empty']) {
 				$result = null;
-			} else {
+			}
+			else {
 				$result = $value['rgb'];
 				// 0..1 float value to hex zero-padded
 				$result .= sprintf('%02s', dechex($value['alfa']));
@@ -199,14 +203,14 @@ class ValueFormater
 			return $value === '-' ? '' : $value;
 		};
 
-		if (isset($formater[$type][$format])) {
+		if(isset($formater[$type][$format])) {
 			return $formater[$type][$format]($value);
-		} else {
+		}
+		else {
 			return $value;
 		}
 	}
-	public static function encode($value, $type, $format)
-	{
+	public static function encode($value, $type, $format) {
 		// Empty values return empty string
 		$formater[self::TYPE_DATE][self::FORMAT_EDITOR] = function ($value) {
 			return $value ? date('Y-m-d', $value) : '';
@@ -216,9 +220,9 @@ class ValueFormater
 				ValueFormater::PERIOD_OF_VALIDITY_START => '',
 				ValueFormater::PERIOD_OF_VALIDITY_END => '',
 			];
-			if (is_array($aValue)) {
+			if(is_array($aValue)) {
 				foreach ($aValue as $attributeName => $value) {
-					if ($value && is_finite($value)) {
+					if($value && is_finite($value)) {
 						$result[$attributeName] = date('Y-m-d', $value);
 					}
 				}
@@ -234,9 +238,9 @@ class ValueFormater
 				ValueFormater::PERIOD_OF_VALIDITY_START => '',
 				ValueFormater::PERIOD_OF_VALIDITY_END => '',
 			];
-			if (is_array($aValue)) {
+			if(is_array($aValue)) {
 				foreach ($aValue as $attributeName => $value) {
-					if ($value && is_finite($value)) {
+					if($value && is_finite($value)) {
 						$result[$attributeName] = date('Y-m-d\TH:i:s\Z', $value);
 					}
 				}
@@ -246,13 +250,13 @@ class ValueFormater
 		};
 		// Purge infinite values in FORMAT_TOKENIZE, transform to empty string (infinite values not accepted in json grammar)
 		$formater[self::TYPE_DATE][self::FORMAT_TOKENIZE] = function ($value) {
-			$result = is_finite((float) $value) ? $value : '';
+			$result = is_finite((double) $value) ? $value : '';
 			return $result;
 		};
 		$formater[self::TYPE_DATE_RANGE][self::FORMAT_TOKENIZE] = function ($aValue) {
 			$result = [];
 			foreach ($aValue as $key => $value) {
-				$result[$key] = is_finite((float) $value) ? $value : '';
+				$result[$key] = is_finite((double) $value) ? $value : '';
 			}
 			return $result;
 		};
@@ -265,9 +269,10 @@ class ValueFormater
 			return $value ? ' checked="checked"' : '';
 		};
 		$formater[self::TYPE_DATE_TIME][self::FORMAT_HUMAN] = function ($value) {
-			if ($value) {
+			if($value) {
 				return date('j M G:i\h', $value);
-			} else {
+			}
+			else {
 				return '';
 			}
 		};
@@ -277,9 +282,9 @@ class ValueFormater
 				ValueFormater::PERIOD_OF_VALIDITY_END => '∞',
 			];
 			$bModified = false;
-			if (is_array($aValue)) {
+			if(is_array($aValue)) {
 				foreach ($aValue as $attributeName => $value) {
-					if ($value && is_finite($value)) {
+					if($value && is_finite($value)) {
 						$result[$attributeName] = date('j M G:i\h', $value);
 						$bModified = true;
 					}
@@ -312,7 +317,7 @@ class ValueFormater
 			//$result = ['rgb' => '#000000'];
 
 			$result['rgb'] = null;
-			if ($value) {
+			if($value) {
 				$result['rgb'] = '#' . substr($value, 1, 6);
 			}
 
@@ -323,7 +328,7 @@ class ValueFormater
 
 			$result['rgb'] = null;
 			$result['alfa'] = null;
-			if ($value) {
+			if($value) {
 				$result['rgb'] = '#' . substr($value, 1, 6);
 				$result['alfa'] = hexdec(substr($value, 7, 2));
 			}
@@ -333,9 +338,10 @@ class ValueFormater
 			return $value ? $value : '-';
 		};
 
-		if (isset($formater[$type][$format])) {
+		if(isset($formater[$type][$format])) {
 			return $formater[$type][$format]($value);
-		} else {
+		}
+		else {
 			return $value;
 		}
 	}
